refactor(diff-dialog): diff_dialog.py(1202줄)를 워커/위젯/서브다이얼로그로 분리
대형 파일 정리 5번째(Phase E). SchemaDiffDialog(760줄)를 뺀 나머지 4개
클래스가 전부 독립적(교차참조 없음, 전부 SchemaDiffDialog가 bare-name으로
생성)이라 db_dialogs.py와 동일한 패턴으로 분리.

- diff_workers.py (신규): SchemaCompareThread, SchemaLoadThread
- diff_pixel_loading_widget.py (신규): PixelLoadingWidget
- diff_sync_script_dialog.py (신규): SyncScriptDialog
- diff_dialog.py: 1202 -> 800줄 (SchemaDiffDialog + 전체 재수입)

logger(get_logger 호출)는 실제로는 SchemaDiffDialog가 아니라
SchemaLoadThread에서만 쓰이고 있어서 diff_workers.py로 함께 이동.

테스트 수정: SchemaLoadThread가 스스로 MySQLConnector를 생성하는 코드가
diff_workers 모듈로 이동했으므로, 해당 흐름을 검증하는 TestLoadSchemas
7곳(공용 dialog fixture 포함)의 monkeypatch 대상을 diff_dialog ->
diff_workers로 이동. start_compare()가 직접 연결하는 다른 경로(diff_dialog.
MySQLConnector patch 6곳)는 그 코드가 계속 facade에 있으므로 변경 없음.

검증: 관련 pytest 32 passed, 전체 pytest 1809 passed / 1 known-flaky 제외.
py_compile 통과. SchemaDiffDialog/SyncScriptDialog 오프스크린 오픈 확인.
앱 스모크 성공.
