+++ src/ui/dialogs/diff_dialog.py
@@ -24,328 +24,15 @@ from src.core.schema_diff import (
     SeverityClassifier, VersionContext, SeveritySummary
 )
 from src.core.db_connector import MySQLConnector
-from src.core.logger import get_logger
-
-logger = get_logger(__name__)
-
-
-class SchemaCompareThread(QThread):
-    """스키마 비교 백그라운드 스레드"""
-
-    progress = pyqtSignal(str)
-    # NOTE: QThread에는 인자 없는 기본 finished 시그널이 있으므로,
-    # 이름을 겹치지 않게 compare_finished로 분리한다.
-    compare_finished = pyqtSignal(list, object, object)  # diffs, SeveritySummary, VersionContext
-    error = pyqtSignal(str)
-
-    def __init__(self, source_connector, target_connector,
-                 source_schema: str, target_schema: str,
-                 compare_level: CompareLevel = CompareLevel.STANDARD):
-        super().__init__()
-        self.source_connector = source_connector
-        self.target_connector = target_connector
-        self.source_schema = source_schema
-        self.target_schema = target_schema
-        self.compare_level = compare_level
-
-    def run(self):
-        try:
-            # MySQL 버전 감지
-            self.progress.emit("MySQL 버전 확인 중...")
-            version_ctx = VersionContext(
-                source_version=self.source_connector.get_db_version(),
-                target_version=self.target_connector.get_db_version(),
-                source_version_str=self.source_connector.get_db_version_string(),
-                target_version_str=self.target_connector.get_db_version_string(),
-            )
-
-            self.progress.emit("소스 스키마 추출 중...")
-            source_extractor = SchemaExtractor(self.source_connector)
-            source_tables = source_extractor.extract_all_tables(self.source_schema)
-
-            self.progress.emit("타겟 스키마 추출 중...")
-            target_extractor = SchemaExtractor(self.target_connector)
-            target_tables = target_extractor.extract_all_tables(self.target_schema)
-
-            self.progress.emit("스키마 비교 중...")
-            comparator = SchemaComparator()
-            diffs = comparator.compare_schemas(
-                source_tables, target_tables, self.compare_level
-            )
-
-            # 심각도 분류
-            self.progress.emit("심각도 분류 중...")
-            classifier = SeverityClassifier(version_ctx)
-            diffs, summary = classifier.classify(diffs)
-
-            self.compare_finished.emit(diffs, summary, version_ctx)
-
-        except Exception as e:
-            self.error.emit(str(e))
+from src.ui.dialogs.diff_workers import SchemaCompareThread, SchemaLoadThread
+from src.ui.dialogs.diff_pixel_loading_widget import PixelLoadingWidget
+from src.ui.dialogs.diff_sync_script_dialog import SyncScriptDialog
 
 
-class SchemaLoadThread(QThread):
-    """스키마 목록 조회 백그라운드 스레드
-
-    DB 연결/조회/해제가 UI 스레드를 블로킹하지 않도록 별도 스레드에서 수행한다.
-    """
-
-    loaded = pyqtSignal(str, list)       # side, schema_names
-    load_failed = pyqtSignal(str, str)   # side, display_message
-
-    def __init__(self, side: str, host: str, port: int,
-                 user: str, password: str):
-        super().__init__()
-        self.side = side
-        self.host = host
-        self.port = port
-        self.user = user
-        self.password = password
-
-    def run(self):
-        connector = None
-        try:
-            connector = MySQLConnector(
-                host=self.host, port=self.port,
-                user=self.user, password=self.password
-            )
-
-            success, _ = connector.connect()
-            if not success:
-                self.load_failed.emit(self.side, "(연결 실패)")
-                return
-
-            schemas = connector.get_schemas(use_cache=False)
-            self.loaded.emit(self.side, list(schemas))
-
-        except Exception as e:
-            logger.error(f"스키마 로드 실패: {e}")
-            self.load_failed.emit(self.side, "(오류)")
-        finally:
-            if connector:
-                try:
-                    connector.disconnect()
-                except Exception:
-                    pass
 
 
-class PixelLoadingWidget(QWidget):
-    """스키마 비교 진행 중 Pixel 아트 애니메이션 위젯
 
-    두 DB 아이콘 사이를 데이터 파티클이 흐르고,
-    활성 단계의 DB가 펄스 효과로 강조된다.
-    """
 
-    TIPS = [
-        "컬럼, 인덱스, FK 정보를 수집하고 있어요",
-        "비교 완료 후 🟡 항목을 클릭하면 상세 내용을 볼 수 있어요",
-        "동기화 스크립트로 타겟 DB 구조를 소스와 맞출 수 있어요",
-        "인덱스와 Foreign Key 변경사항도 자동 감지합니다",
-        "행 수(row count) 차이도 함께 비교합니다",
-    ]
-
-    PX = 4  # 1 pixel art pixel = 4×4 real pixels
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setFixedHeight(90)
-        self.setVisible(False)
-
-        self._phase = "idle"   # idle / source / target / compare
-        self._frame = 0
-        self._tip_idx = 0
-        self._status_text = ""
-        self._particles: list[dict] = []
-
-        self._anim_timer = QTimer(self)
-        self._anim_timer.timeout.connect(self._tick)
-        self._tip_timer = QTimer(self)
-        self._tip_timer.timeout.connect(self._next_tip)
-
-    # ------ public API ------
-
-    def start(self, status: str):
-        """애니메이션 시작"""
-        self._status_text = status
-        self._phase = self._detect_phase(status)
-        self._frame = 0
-        self._tip_idx = 0
-        self._init_particles()
-        self.setVisible(True)
-        self._anim_timer.start(80)   # ~12 FPS
-        self._tip_timer.start(3500)
-        self.update()
-
-    def update_status(self, status: str):
-        """진행 상태(phase) 업데이트"""
-        self._status_text = status
-        new_phase = self._detect_phase(status)
-        if new_phase != self._phase:
-            self._phase = new_phase
-            self._init_particles()
-        self.update()
-
-    def stop(self):
-        """애니메이션 정지 및 숨김"""
-        self._anim_timer.stop()
-        self._tip_timer.stop()
-        self.setVisible(False)
-        self._phase = "idle"
-
-    # ------ internals ------
-
-    @staticmethod
-    def _detect_phase(s: str) -> str:
-        if "소스" in s:
-            return "source"
-        if "타겟" in s:
-            return "target"
-        if "비교" in s:
-            return "compare"
-        return "source"
-
-    def _tick(self):
-        self._frame += 1
-        self._update_particles()
-        self.update()
-
-    def _next_tip(self):
-        self._tip_idx = (self._tip_idx + 1) % len(self.TIPS)
-
-    def _init_particles(self):
-        self._particles = []
-        n = 8 if self._phase == "compare" else 5
-        for _ in range(n):
-            self._particles.append({
-                "x": random.uniform(0, 1),
-                "y": random.uniform(0.2, 0.8),
-                "speed": random.uniform(0.012, 0.028),
-                "size": random.choice([1, 2]),
-                "alt": random.random() > 0.5,
-            })
-
-    def _update_particles(self):
-        for i, p in enumerate(self._particles):
-            p["x"] += p["speed"]
-            if p["x"] > 1.0:
-                self._particles[i] = {
-                    "x": 0.0,
-                    "y": random.uniform(0.2, 0.8),
-                    "speed": random.uniform(0.012, 0.028),
-                    "size": random.choice([1, 2]),
-                    "alt": random.random() > 0.5,
-                }
-
-    # ------ painting ------
-
-    def paintEvent(self, event):  # noqa: N802
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.RenderHint.Antialiasing, False)
-
-        w = self.width()
-        px = self.PX
-        cx = w // 2
-
-        # DB icon positions (중앙 대칭 배치, DB 아이콘 너비 = 7*px = 28px)
-        db_icon_w = 7 * px      # 28px
-        half_gap = 52           # 중앙에서 각 아이콘까지의 거리
-        src_x = cx - half_gap - db_icon_w   # 소스: 중앙 왼쪽
-        tgt_x = cx + half_gap               # 타겟: 중앙 오른쪽
-        db_y = 2
-
-        # Source DB
-        src_active = self._phase in ("source", "compare")
-        src_c = QColor("#3498db") if src_active else QColor("#bdc3c7")
-        self._draw_db(painter, src_x, db_y, px, src_c,
-                       pulse=(self._phase == "source"))
-
-        # Target DB
-        tgt_active = self._phase in ("target", "compare")
-        tgt_c = QColor("#2ecc71") if tgt_active else QColor("#bdc3c7")
-        self._draw_db(painter, tgt_x, db_y, px, tgt_c,
-                       pulse=(self._phase == "target"))
-
-        # Labels
-        painter.setPen(QPen(QColor("#7f8c8d")))
-        painter.setFont(QFont("Consolas", 7))
-        label_y = db_y + 9 * px + 2
-        painter.drawText(src_x, label_y, 7 * px, 12,
-                         Qt.AlignmentFlag.AlignCenter, "SRC")
-        painter.drawText(tgt_x, label_y, 7 * px, 12,
-                         Qt.AlignmentFlag.AlignCenter, "TGT")
-
-        # Dotted connection line
-        line_left = src_x + 8 * px
-        line_right = tgt_x - px
-        line_y = db_y + 4 * px
-        painter.setPen(QPen(QColor("#dcdde1"), 1, Qt.PenStyle.DotLine))
-        painter.drawLine(line_left, line_y, line_right, line_y)
-
-        # Flowing data particles
-        gap = line_right - line_left
-        if gap > 0:
-            for p in self._particles:
-                px_x = int(line_left + p["x"] * gap)
-                px_y = int(line_y - 6 + p["y"] * 12)
-                c = QColor("#f39c12") if not p["alt"] else QColor("#e74c3c")
-                s = p["size"] * px
-                painter.fillRect(px_x, px_y, s, s, c)
-
-        # Compare phase: pulsing center indicator
-        if self._phase == "compare":
-            pulse = abs(math.sin(self._frame * 0.15))
-            size = int(4 + pulse * 6)
-            c = QColor("#9b59b6")
-            c.setAlpha(int(120 + pulse * 135))
-            painter.fillRect(cx - size // 2, line_y - size // 2, size, size, c)
-
-        # Status text with animated dots
-        dots = "." * ((self._frame // 4) % 4)
-        text = self._status_text.rstrip(".") + dots
-        painter.setPen(QPen(QColor("#2c3e50")))
-        painter.setFont(QFont("Consolas", 10, QFont.Weight.Bold))
-        painter.drawText(0, 52, w, 18, Qt.AlignmentFlag.AlignCenter, text)
-
-        # Rotating tip
-        painter.setPen(QPen(QColor("#95a5a6")))
-        painter.setFont(QFont("", 9))
-        painter.drawText(0, 72, w, 16, Qt.AlignmentFlag.AlignCenter,
-                         self.TIPS[self._tip_idx])
-
-        painter.end()
-
-    def _draw_db(self, painter: QPainter, x: int, y: int,
-                 px: int, color: QColor, pulse: bool = False):
-        """Pixel art database cylinder (7×9 art pixels)"""
-        light = QColor(color).lighter(130)
-        dark = QColor(color).darker(130)
-        fill = QColor("#f8f9fa")
-
-        # Pulse glow
-        if pulse and (self._frame // 6) % 2 == 0:
-            g = QColor(color)
-            g.setAlpha(35)
-            painter.fillRect(x - px, y - px, 9 * px, 11 * px, g)
-
-        # Top cap
-        painter.fillRect(x + 2 * px, y, 3 * px, px, light)
-        painter.fillRect(x + px, y + px, 5 * px, px, color)
-
-        # Side walls
-        for r in range(2, 8):
-            painter.fillRect(x, y + r * px, px, px, color)
-            painter.fillRect(x + 6 * px, y + r * px, px, px, color)
-
-        # Interior fill (two sections)
-        painter.fillRect(x + px, y + 2 * px, 5 * px, 2 * px, fill)
-        painter.fillRect(x + px, y + 5 * px, 5 * px, 2 * px, fill)
-
-        # Middle divider
-        painter.fillRect(x, y + 4 * px, 7 * px, px, light)
-
-        # Bottom
-        painter.fillRect(x + px, y + 7 * px, 5 * px, px, color)
-        painter.fillRect(x + 2 * px, y + 8 * px, 3 * px, px, dark)
 
 
 class SchemaDiffDialog(QDialog):
@@ -1111,92 +798,3 @@ class SchemaDiffDialog(QDialog):
                 thread.wait(3000)
 
 
-class SyncScriptDialog(QDialog):
-    """동기화 스크립트 다이얼로그"""
-
-    def __init__(self, parent=None, script: str = ""):
-        super().__init__(parent)
-        self.script = script
-        self._setup_ui()
-
-    def _setup_ui(self):
-        """UI 구성"""
-        self.setWindowTitle("동기화 스크립트")
-        self.setMinimumSize(700, 500)
-
-        layout = QVBoxLayout(self)
-
-        # 경고
-        warning = QLabel(
-            "⚠️ 주의: 이 스크립트를 실행하기 전에 반드시 타겟 데이터베이스를 백업하세요!"
-        )
-        warning.setStyleSheet(
-            "background-color: #fff3cd; color: #856404; "
-            "padding: 10px; border-radius: 4px; font-weight: bold;"
-        )
-        warning.setWordWrap(True)
-        layout.addWidget(warning)
-
-        # 데이터 미복사 경고
-        data_warning = QLabel(
-            "📋 이 스크립트는 테이블 구조(DDL)만 동기화합니다.\n"
-            "데이터는 복사되지 않습니다. 데이터 이전은 Export/Import 기능을 사용하세요."
-        )
-        data_warning.setStyleSheet(
-            "background-color: #d1ecf1; color: #0c5460; "
-            "padding: 10px; border-radius: 4px; font-weight: bold;"
-        )
-        data_warning.setWordWrap(True)
-        layout.addWidget(data_warning)
-
-        # 스크립트
-        self.script_text = QTextEdit()
-        self.script_text.setPlainText(self.script)
-        self.script_text.setFont(QFont("Consolas", 10))
-        self.script_text.setReadOnly(True)
-        layout.addWidget(self.script_text)
-
-        # 버튼
-        btn_layout = QHBoxLayout()
-
-        copy_btn = QPushButton("클립보드에 복사")
-        copy_btn.clicked.connect(self._copy_to_clipboard)
-        btn_layout.addWidget(copy_btn)
-
-        save_btn = QPushButton("파일로 저장")
-        save_btn.clicked.connect(self._save_to_file)
-        btn_layout.addWidget(save_btn)
-
-        btn_layout.addStretch()
-
-        close_btn = QPushButton("닫기")
-        close_btn.clicked.connect(self.accept)
-        btn_layout.addWidget(close_btn)
-
-        layout.addLayout(btn_layout)
-
-    def _copy_to_clipboard(self):
-        """클립보드에 복사"""
-        clipboard = QApplication.clipboard()
-        clipboard.setText(self.script)
-        QMessageBox.information(self, "복사 완료", "스크립트가 클립보드에 복사되었습니다.")
-
-    def _save_to_file(self):
-        """파일로 저장"""
-        file_path, _ = QFileDialog.getSaveFileName(
-            self,
-            "스크립트 저장",
-            "sync_script.sql",
-            "SQL Files (*.sql);;All Files (*)"
-        )
-
-        if file_path:
-            try:
-                with open(file_path, 'w', encoding='utf-8') as f:
-                    f.write(self.script)
-                QMessageBox.information(
-                    self, "저장 완료",
-                    f"스크립트가 저장되었습니다:\n{file_path}"
-                )
-            except Exception as e:
-                QMessageBox.critical(self, "저장 실패", f"파일 저장 실패: {e}")
